Update package for java bindings.
