+++ java/src/main/java/ai/rapids/kvikio/cufile/CuFileReadHandle.java
@@ -14,26 +14,21 @@
  * limitations under the License.
  */
 
-package bindings.kvikio.cufile;
+package ai.rapids.kvikio.cufile;
 
-abstract class CuFileHandle implements AutoCloseable {
-    private final long pointer;
+public final class CuFileReadHandle extends CuFileHandle {
 
-    static {
-        CuFile.initialize();
+    public CuFileReadHandle(String path) {
+        super(create(path));
     }
 
-    protected CuFileHandle(long pointer) {
-        this.pointer = pointer;
+    public void read(long device_pointer, long size, long file_offset, long device_offset) {
+        readFile(getPointer(), device_pointer, size, file_offset, device_offset);
     }
 
-    public void close() {
-        destroy(pointer);
-    }
+    private static native long create(String path);
 
-    protected long getPointer() {
-        return this.pointer;
-    }
+    private static native void readFile(long file_pointer, long device_pointer, long size, long file_offset,
+            long device_offset);
 
-    private static native void destroy(long pointer);
 }
